requirements.txt
flask>=3.0
gunicorn>=21.2
stripe>=8.0
resend>=2.0
apscheduler>=3.10,<4.0
sqlalchemy>=2.0
python-dotenv>=1.0
